requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.6
Django==6.0.3
idna==3.11
newsapi-python==0.2.7
pillow==12.1.1
python-dotenv==1.2.2
requests==2.33.0
sqlparse==0.5.5
urllib3==2.6.3
